requirements.txt
# Core scraping dependencies
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# UI Framework
streamlit>=1.28.0
plotly>=5.17.0

# Configuration and utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and monitoring
loguru>=0.7.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0